Refactor name project to match standard pattern, refactor some bad practices in project, and add some dependencies for future use
