--- src/main/java/com/stagproj/BankingAPI/BankApplication.java
@@ -1,15 +1,0 @@
-package com.stagproj.BankingAPI;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication
-@EnableJpaRepositories("com.stagproj.BankingAPI.repositories")
-public class BankApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BankApplication.class, args);
-	}
-
-}
