--- src/main/java/com/stagproj/BankingAPI/core/ApiErrors.java
@@ -1,19 +1,0 @@
-package com.stagproj.BankingAPI.core;
-
-import lombok.Getter;
-
-import java.util.Collections;
-import java.util.List;
-
-public class ApiErrors {
-    @Getter
-    private final List<String> errors;
-
-    public ApiErrors(List<String> errors) {
-        this.errors = errors;
-    }
-
-    public ApiErrors(String mensagem){
-        this.errors = Collections.singletonList(mensagem);
-    }
-}
