--- src/main/java/com/stagproj/BankingAPI/dtos/reponses/PessoaResponse.java
@@ -1,23 +1,0 @@
-package com.stagproj.BankingAPI.dtos.reponses;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDate;
-
-@Getter
-@Setter
-public class PessoaResponse {
-
-    private long id;
-
-    private String nome;
-
-    private String cpf;
-
-    @JsonFormat(pattern="dd-MM-yyyy")
-    private LocalDate dataNascimento;
-
-    private long idConta;
-}
