--- src/main/java/com/stagproj/BankingAPI/entities/Transacao.java
@@ -1,32 +1,0 @@
-package com.stagproj.BankingAPI.entities;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Transacao {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NonNull
-    private double valor;
-
-    @NonNull
-    private LocalDate dataTransacao;
-
-    @ManyToOne
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    @JoinColumn(name = "idConta", referencedColumnName = "id")
-    private Conta conta;
-
-}
