--- src/main/java/com/stagproj/BankingAPI/exceptions/ExceptionMethod.java
@@ -1,8 +1,0 @@
-package com.stagproj.BankingAPI.exceptions;
-
-public class ExceptionMethod extends RuntimeException {
-
-    public ExceptionMethod(String message){
-        super(message);
-    }
-}
