--- src/main/java/com/stagproj/BankingAPI/repositories/ContaRepository.java
@@ -1,9 +1,0 @@
-package com.stagproj.BankingAPI.repositories;
-
-import com.stagproj.BankingAPI.entities.Conta;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ContaRepository extends JpaRepository<Conta, Long> {
-}
