+++ src/main/java/com/stagproj/banking/BankApplication.java
@@ -1,0 +1,15 @@
+package com.stagproj.banking;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EnableJpaRepositories("com.stagproj.banking.repositories")
+public class BankApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BankApplication.class, args);
+	}
+
+}
