+++ src/main/java/com/stagproj/banking/core/ModelMapperConfig.java
@@ -1,0 +1,14 @@
+package com.stagproj.banking.core;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+}
