+++ src/main/java/com/stagproj/banking/exceptions/ExceptionMethod.java
@@ -1,0 +1,8 @@
+package com.stagproj.banking.exceptions;
+
+public class ExceptionMethod extends RuntimeException {
+
+    public ExceptionMethod(String message){
+        super(message);
+    }
+}
