+++ src/main/java/com/stagproj/banking/repositories/ContaRepository.java
@@ -1,0 +1,9 @@
+package com.stagproj.banking.repositories;
+
+import com.stagproj.banking.entities.Conta;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ContaRepository extends JpaRepository<Conta, Long> {
+}
