+++ src/main/java/com/stagproj/banking/services/ContaService.java
@@ -1,0 +1,13 @@
+package com.stagproj.banking.services;
+import com.stagproj.banking.entities.Conta;
+import java.util.List;
+
+public interface ContaService {
+    List<Conta> getAll();
+
+    Conta getConta(long id);
+
+    Conta criacaoConta(Conta conta);
+
+    Conta bloqueioConta(long id);
+}
