+++ src/main/java/com/stagproj/banking/services/TransacaoService.java
@@ -1,0 +1,11 @@
+package com.stagproj.banking.services;
+import com.stagproj.banking.entities.Transacao;
+import java.util.List;
+
+public interface TransacaoService {
+    Transacao deposito(Transacao transacao);
+
+    Transacao saque(Transacao transacao);
+
+    List<Transacao> extrato(long id);
+}
